Small tail

+++ build.sh
@@ -7,27 +7,30 @@ TEST_TARGET_DIR="${BUILD_DIR}/test/${TEST_TARGET}"
 PROFILE_FILE="${PWD}/${BUILD_DIR}/coverage-%p.profraw"
 MERGED_FILE="${BUILD_DIR}/coverage.profdata"
 HTML_DIR="${BUILD_DIR}/coverage_html"
+TARGET_TYPE="Release"
 
-cmake -S . -B "$BUILD_DIR" -DENABLE_COVERAGE=ON -DCMAKE_BUILD_TYPE=Debug \
+cmake -S . -B "$BUILD_DIR" -DENABLE_COVERAGE=ON -DCMAKE_BUILD_TYPE=${TARGET_TYPE} \
     -DCMAKE_C_COMPILER=/opt/homebrew/opt/llvm/bin/clang \
     -DCMAKE_CXX_COMPILER=/opt/homebrew/opt/llvm/bin/clang++
 
 cmake --build "$BUILD_DIR" -j8
 
-export LLVM_PROFILE_FILE="$PROFILE_FILE"
-cd "$BUILD_DIR"
-ctest --output-on-failure
-cd ..
+if [ $TARGET_TYPE = "Debug" ]; then
+    export LLVM_PROFILE_FILE="$PROFILE_FILE"
+    cd "$BUILD_DIR"
+    ctest --output-on-failure
+    cd ..
 
-rm -rf ${MERGED_FILE}
-llvm-profdata merge -sparse "$BUILD_DIR"/coverage-*.profraw -o "$MERGED_FILE"
-rm -rf ${BUILD_DIR}/*.profraw
+    rm -rf ${MERGED_FILE}
+    llvm-profdata merge -sparse "$BUILD_DIR"/coverage-*.profraw -o "$MERGED_FILE"
+    rm -rf ${BUILD_DIR}/*.profraw
 
-llvm-cov show "$TEST_TARGET_DIR" \
-    -instr-profile="$MERGED_FILE" \
-    -format=html \
-    -output-dir="$HTML_DIR" \
-    --show-branches=percent \
-    $(find "$PWD/plugin/source" -name '*.cpp' -o -name '*.h')
+    llvm-cov show "$TEST_TARGET_DIR" \
+        -instr-profile="$MERGED_FILE" \
+        -format=html \
+        -output-dir="$HTML_DIR" \
+        --show-branches=percent \
+        $(find "$PWD/plugin/source" -name '*.cpp' -o -name '*.h')
 
-echo "Coverage report generated at $HTML_DIR/index.html"
+    echo "Coverage report generated at $HTML_DIR/index.html"
+fi
